Move Dribbble parsing to encodingconverter
